Re-organize error handling code
